A large number of changes mostly related to the adding of new patients.
The system now is able to use the old method and add pateints to the database. This is now done with a dialogue with validation instead of displaying static text boxes on the main screen.

The daabase is queried and all currently included patients are now able to be shown in the patient screen's main display, with a datagrid being populated using binding and set column headers.

There are a number of changes still to be made, but this is a step in the right direction